requirements.txt
pdfminer.six
nltk
scikit-learn
